igblastr 0.2.2
All built-in C-region dbs (and upcoming built-in germline dbs) are
now versioned. The version is the date (in YYYYMM format) when the
corresponding FASTA files were downloaded and added to the igblastr
package (under inst/extdata/constant_regions/ for the built-in
C-region dbs and under inst/extdata/germline_sequences/ for the
built-in germline dbs).